+++ app/components/signing/processing_component.html.erb
@@ -11,24 +11,24 @@
     <div class="py-4">
       <%= content_tag(:div,
                       {
-                        "data-controller": "new-autogram",
-                        "data-new-autogram-files-to-be-signed": files_to_be_signed.to_json.html_safe,
-                        "data-new-autogram-authenticity-token": form_authenticity_token
+                        "data-controller": "autogram",
+                        "data-autogram-files-to-be-signed": files_to_be_signed.to_json.html_safe,
+                        "data-autogram-authenticity-token": form_authenticity_token
                       }
           ) do %>
-        <div data-new-autogram-target="appNotRunning">
-          <%= link_to "Autogram", "autogram://go", class: "hidden", data: { "new-autogram-target": "openAutogramApp" } %>
+        <div data-autogram-target="appNotRunning">
+          <%= link_to "Autogram", "autogram://go", class: "hidden", data: { "autogram-target": "openAutogramApp" } %>
           <p class="pb-4">
             Spustite aplikáciu
             <%= link_to "Autogram", "https://sluzby.slovensko.digital/autogram/", target: "_blank", class: "underline font-bold hover:no-underline" %>
             a skúste znova
           </p>
-          <div data-action="click->new-autogram#signAll" class="inline-flex w-full justify-center items-center px-3 py-2 rounded-md bg-blue-600 font-medium text-left text-sm text-white hover:cursor-pointer sm:w-auto">
+          <div data-action="click->autogram#signAll" class="inline-flex w-full justify-center items-center px-3 py-2 rounded-md bg-blue-600 font-medium text-left text-sm text-white hover:cursor-pointer sm:w-auto">
             Skúsiť znova
           </div>
         </div>
 
-        <div data-new-autogram-target="signingInProgress">
+        <div data-autogram-target="signingInProgress">
           <p class="pb-4">Prosím, nezatvárajte okno prehliadača.</p>
         </div>
 
@@ -36,14 +36,14 @@
                       @after_singing_path,
                       method: :patch,
                       params: { result: "ok" },
-                      data: { "new-autogram-target": "doneOk" },
+                      data: { "autogram-target": "doneOk" },
                       class: "hidden" %>
 
         <%= button_to "Podpisovanie zlyhalo",
                       @after_singing_path,
                       method: :patch,
                       params: { result: "error" },
-                      data: { "new-autogram-target": "doneError" },
+                      data: { "autogram-target": "doneError" },
                       class: "hidden" %>
       <% end %>
     </div>
