+++ app/javascript/controllers/autogram_controller.js
